pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "Starter template for making a new agent LangGraph."
authors = [
    { name = "William Fu-Hinthorn", email = "13333726+hinthornw@users.noreply.github.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<3.13"
dependencies = [
    "langgraph==0.6.2",
    "langgraph-prebuilt==0.6.2",
    "langgraph-checkpoint==2.1.1",
    "langgraph-sdk==0.2.0",
    "langchain-core==0.3.72",
    "langchain==0.3.27",
    "langchain-text-splitters==0.3.9",
    "langchain-google-vertexai==2.0.28",
    "langchain-mcp-adapters (>=0.1.9,<0.2.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "starlette==0.47.2",
    "uvicorn==0.35.0",
    "sse-starlette==2.1.3",
    "a2a-sdk @ git+https://github.com/a2aproject/a2a-python.git"
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "langgraph-cli[inmem]==0.3.6",
    "langgraph-api==0.2.102",
    "langgraph-runtime-inmem==0.6.1",
    "pytest>=8.3.5",
    "mypy>=1.13.0",
    "ruff>=0.8.2",
    "anyio>=4.7.0",
]
